+++ packages/create-sitecore-jss/src/templates/nextjs-sxa/scripts/config/plugins/sxa.ts
@@ -1,7 +1,8 @@
-import { ConfigPlugin, JssConfig } from '..';
+import { JssConfig } from 'lib/config';
+import { ConfigPlugin } from '..';
 
 /**
- * This plugin will set configuration specific for SXA. 
+ * This plugin will set configuration specific for SXA.
  */
 class SXAPlugin implements ConfigPlugin {
   // should come before fallback
