+++ packages/create-sitecore-jss/src/templates/nextjs/.env
@@ -21,10 +21,7 @@ JSS_EDITING_SECRET=
 
 # ===== Sitecore Edge Platform ======
 
-# Your Sitecore Edge Platform URL.
-SITECORE_EDGE_URL=https://edge-platform.sitecorecloud.io
-
-# Your unified Sitecore Edge Context Id is needed to build the app.
+# Your unified Sitecore Edge Context Id.
 SITECORE_EDGE_CONTEXT_ID=
 
 # ====== Sitecore Preview / Delivery Edge ======
