+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/config/index.ts
@@ -1,20 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const plugins = require('scripts/temp/config-plugins');
-
-/**
- * JSS configuration object
- */
-export interface JssConfig extends Record<string, string | undefined> {
-  sitecoreApiKey?: string;
-  sitecoreApiHost?: string;
-  sitecoreEdgeUrl?: string;
-  sitecoreEdgeContextId?: string;
-  siteName?: string;
-  graphQLEndpointPath?: string;
-  defaultLanguage?: string;
-  graphQLEndpoint?: string;
-  layoutServiceConfigurationName?: string;
-}
+import { JssConfig } from 'lib/config';
 
 export interface ConfigPlugin {
   /**
