+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/config/plugins/computed.ts
@@ -1,4 +1,5 @@
-import { ConfigPlugin, JssConfig } from '..';
+import { JssConfig } from 'lib/config';
+import { ConfigPlugin } from '..';
 
 /**
  * This plugin will set computed config props.
