+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/config/plugins/fallback.ts
@@ -1,5 +1,6 @@
 import chalk from 'chalk';
-import { ConfigPlugin, JssConfig } from '..';
+import { JssConfig } from 'lib/config';
+import { ConfigPlugin } from '..';
 
 /**
  * This config will set fallback values for properties that were left empty
@@ -22,6 +23,7 @@ class FallbackPlugin implements ConfigPlugin {
       defaultLanguage: config.defaultLanguage || 'en',
       sitecoreApiKey: config.sitecoreApiKey || 'no-api-key-set',
       layoutServiceConfigurationName: config.layoutServiceConfigurationName || 'default',
+      sitecoreEdgeUrl: config.sitecoreEdgeUrl || 'https://edge-platform.sitecorecloud.io',
     });
   }
 }
