+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/config/plugins/package-json.ts
@@ -1,4 +1,5 @@
-import { ConfigPlugin, JssConfig } from '..';
+import { JssConfig } from 'lib/config';
+import { ConfigPlugin } from '..';
 import packageConfig from 'package.json';
 
 /**
