+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/config/plugins/scjssconfig.ts
@@ -1,4 +1,5 @@
-import { ConfigPlugin, JssConfig } from '..';
+import { JssConfig } from 'lib/config';
+import { ConfigPlugin } from '..';
 
 /**
  * This plugin will set config props based on scjssconfig.json.
