+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/generate-config.ts
@@ -2,7 +2,8 @@ import 'dotenv/config';
 import fs from 'fs';
 import path from 'path';
 import { constantCase } from 'constant-case';
-import { JssConfig, jssConfigFactory } from './config';
+import { JssConfig } from 'lib/config';
+import { jssConfigFactory } from './config';
 
 /*
   CONFIG GENERATION
@@ -15,7 +16,8 @@ const defaultConfig: JssConfig = {
   sitecoreApiHost: process.env[`${constantCase('sitecoreApiHost')}`],
   sitecoreEdgeUrl: process.env[`${constantCase('sitecoreEdgeUrl')}`],
   sitecoreEdgeContextId: process.env[`${constantCase('sitecoreEdgeContextId')}`],
-  siteName: process.env[`${constantCase('siteName')}`],
+  siteName:
+    process.env[`${constantCase('siteName')}`] || process.env[`${constantCase('jssAppName')}`],
   graphQLEndpointPath: process.env[`${constantCase('graphQLEndpointPath')}`],
   defaultLanguage: process.env[`${constantCase('defaultLanguage')}`],
   graphQLEndpoint: process.env[`${constantCase('graphQLEndpoint')}`],
