+++ packages/sitecore-jss-nextjs/src/index.ts
@@ -9,6 +9,7 @@ export {
   NativeDataFetcher,
   NativeDataFetcherConfig,
   GraphQLRequestClient,
+  GraphQLRequestClientFactory,
   GraphQLRequestClientFactoryConfig,
   HTMLLink,
   enableDebug,
