+++ packages/sitecore-jss/src/graphql-request-client.ts
@@ -95,12 +95,15 @@ export class GraphQLRequestClient implements GraphQLClient {
   }
 
   /**
-   * Factory method for creating a GraphQLRequestClient.
+   * Factory method for creating a GraphQLRequestClientFactory.
    * @param {Object} config - client configuration options.
    * @param {string} config.endpoint - endpoint
    * @param {string} [config.apiKey] - apikey
    */
-  static createClientFactory({ endpoint, apiKey }: GraphQLRequestClientFactoryConfig) {
+  static createClientFactory({
+    endpoint,
+    apiKey,
+  }: GraphQLRequestClientFactoryConfig): GraphQLRequestClientFactory {
     return (config: Omit<GraphQLRequestClientConfig, 'apiKey'> = {}) =>
       new GraphQLRequestClient(endpoint, { ...config, apiKey });
   }
