+++ packages/sitecore-jss/src/graphql/graphql-edge-proxy.test.ts
@@ -5,13 +5,23 @@ import { getEdgeProxyContentUrl } from './graphql-edge-proxy';
 describe('graphql-edge-proxy', () => {
   describe('getEdgeProxyContentUrl', () => {
     it('should return url', () => {
-      const endpoint = 'https://edge.staging';
-      const contextId = '0730fc5a-3333-5555-5555-08db6d7ddb49';
+      const sitecoreEdgeContextId = '0730fc5a-3333-5555-5555-08db6d7ddb49';
 
-      const url = getEdgeProxyContentUrl(endpoint, contextId);
+      const url = getEdgeProxyContentUrl(sitecoreEdgeContextId);
 
       expect(url).to.equal(
-        'https://edge.staging/content/api/graphql/v1?sitecoreContextId=0730fc5a-3333-5555-5555-08db6d7ddb49'
+        'https://edge-platform.sitecorecloud.io/content/api/graphql/v1?sitecoreContextId=0730fc5a-3333-5555-5555-08db6d7ddb49'
+      );
+    });
+
+    it('should return url when custom sitecoreEdgeUrl is provided', () => {
+      const sitecoreEdgeUrl = 'https://test.com';
+      const sitecoreEdgeContextId = '0730fc5a-3333-5555-5555-08db6d7ddb49';
+
+      const url = getEdgeProxyContentUrl(sitecoreEdgeContextId, sitecoreEdgeUrl);
+
+      expect(url).to.equal(
+        'https://test.com/content/api/graphql/v1?sitecoreContextId=0730fc5a-3333-5555-5555-08db6d7ddb49'
       );
     });
   });
