+++ packages/sitecore-jss/src/graphql/graphql-edge-proxy.ts
@@ -1,8 +1,10 @@
 /**
  * Generates a URL for accessing Sitecore Edge Platform Content using the provided endpoint and context ID.
- * @param {string} endpoint - The base endpoint URL for the Edge Platform.
- * @param {string} contextId - The unique context ID.
+ * @param {string} sitecoreEdgeContextId - The unique context id.
+ * @param {string} [sitecoreEdgeUrl] - The base endpoint URL for the Edge Platform. Default is https://edge-platform.sitecorecloud.io
  * @returns {string} The complete URL for accessing content through the Edge Platform.
  */
-export const getEdgeProxyContentUrl = (endpoint: string, contextId: string) =>
-  `${endpoint}/content/api/graphql/v1?sitecoreContextId=${contextId}`;
+export const getEdgeProxyContentUrl = (
+  sitecoreEdgeContextId: string,
+  sitecoreEdgeUrl = 'https://edge-platform.sitecorecloud.io'
+) => `${sitecoreEdgeUrl}/content/api/graphql/v1?sitecoreContextId=${sitecoreEdgeContextId}`;
